initialize lightMap with memset

+++ worldmap.cpp
@@ -93,7 +93,8 @@ void WorldMap::Circle(
     }
     float maxs[360] = { (float)qMax(min_rad, float(qrand()%qRound(max_rad))) };
     for (ushort x=1; x<360; ++x) {
-        maxs[x]= qBound(min_rad, maxs[x-1]+(qrand()%400-200.0)/200.0, max_rad);
+        maxs[x]= qBound(
+            min_rad, float(maxs[x-1]+(qrand()%400-200.0)/200.0),max_rad);
         if ( x > 315 ) { // connect beginning and end of circle
             maxs[x] += (maxs[0]-maxs[x-1])/90;
         }
